+++ src/bpp/models/patent.py
@@ -1,4 +1,3 @@
-# -*- encoding: utf-8 -*-
 from denorm import denormalized, depend_on_related
 from django.db import models
 from django.db.models import CASCADE, SET_NULL
@@ -133,6 +132,8 @@ class Patent(
     @depend_on_related(
         "bpp.Patent_Autor",
         only=(
+            "autor_id",
+            "jednostka_id",
             "typ_odpowiedzialnosci_id",
             "afiliuje",
             "dyscyplina_naukowa_id",
@@ -166,8 +167,7 @@ class Patent(
     @depend_on_related("bpp.Patent_Autor", only=("kolejnosc",))
     def opis_bibliograficzny_autorzy_cache(self):
         return [
-            "%s %s" % (x.autor.nazwisko, x.autor.imiona)
-            for x in self.autorzy_dla_opisu()
+            f"{x.autor.nazwisko} {x.autor.imiona}" for x in self.autorzy_dla_opisu()
         ]
 
     @denormalized(models.TextField, blank=True, null=True)