+++ src/bpp/models/wydawnictwo_zwarte.py
@@ -1,4 +1,3 @@
-# -*- encoding: utf-8 -*-
 import re
 import warnings
 
@@ -248,6 +247,8 @@ class Wydawnictwo_Zwarte(
     @depend_on_related(
         "bpp.Wydawnictwo_Zwarte_Autor",
         only=(
+            "autor_id",
+            "jednostka_id",
             "typ_odpowiedzialnosci_id",
             "afiliuje",
             "dyscyplina_naukowa_id",
@@ -286,8 +287,7 @@ class Wydawnictwo_Zwarte(
     @depend_on_related("bpp.Wydawnictwo_Zwarte_Autor", only=("kolejnosc",))
     def opis_bibliograficzny_autorzy_cache(self):
         return [
-            "%s %s" % (x.autor.nazwisko, x.autor.imiona)
-            for x in self.autorzy_dla_opisu()
+            f"{x.autor.nazwisko} {x.autor.imiona}" for x in self.autorzy_dla_opisu()
         ]
 
     @denormalized(models.TextField, blank=True, null=True)