+++ src/ewaluacja2021/management/commands/odepnij_dyscypliny.py
@@ -23,6 +23,7 @@ class Command(BaseCommand):
         with transaction.atomic():
             for rekord in rekordy_danych.exclude(do_ewaluacji=True).exclude(slot=1):
                 original = rekord.rekord.original
+                # Odpinaj tylko w pracach, gdzie jest więcej jak jeden autor:
                 if original.autorzy_set.count() > 1:
                     original.autorzy_set.filter(
                         autor=rekord.autor, dyscyplina_naukowa=rekord.dyscyplina