+++ src/ewaluacja2021/management/commands/raport_3n_plecakowy.py
@@ -7,12 +7,17 @@ from ewaluacja2021.core.plecakowy import Plecakowy
 class Command(BaseCommand):
     def add_arguments(self, parser):
         parser.add_argument("--dyscyplina", default="nauki medyczne")
+        parser.add_argument("--bez-limitu-uczelni", type=bool, default=False)
         parser.add_argument("--output-path", default=None, type=str)
 
     @transaction.atomic
-    def handle(self, dyscyplina, output_path, liczba_n=None, *args, **options):
+    def handle(self, dyscyplina, output_path, bez_limitu_uczelni, *args, **options):
 
-        algorytm = Plecakowy(nazwa_dyscypliny=dyscyplina, output_path=output_path)
+        algorytm = Plecakowy(
+            nazwa_dyscypliny=dyscyplina,
+            bez_limitu_uczelni=bez_limitu_uczelni,
+            output_path=output_path,
+        )
         algorytm.powitanie()
         algorytm.sumuj()
         algorytm.zrzuc_dane("plecakowy")